Changed project scheme name to Severenity, added launch screen, added custom navigation bar view from .xib file
