+++ iOS/Severenity/Views/MapViewController.swift
@@ -86,32 +86,3 @@ class MapViewController: UIViewController, CLLocationManagerDelegate, MapPresent
         }
     }
 }
-
-extension UIImage {
-    
-    func roundedImageWithBorder(with width: CGFloat, and color: UIColor) -> UIImage? {
-        let square = CGSize(width: min(size.width, size.height) + width * 2, height: min(size.width, size.height) + width * 2)
-        let imageView = UIImageView(frame: CGRect(origin: CGPoint(x: 0, y: 0), size: square))
-        imageView.contentMode = .center
-        imageView.image = self
-        imageView.layer.cornerRadius = square.width/2
-        imageView.layer.masksToBounds = true
-        imageView.layer.borderWidth = width
-        imageView.layer.borderColor = color.cgColor
-        UIGraphicsBeginImageContextWithOptions(imageView.bounds.size, false, scale)
-        guard let context = UIGraphicsGetCurrentContext() else { return nil }
-        imageView.layer.render(in: context)
-        let result = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return result
-    }
-    
-    func imageResize (sizeChange:CGSize)-> UIImage{
-        let hasAlpha = true
-        let scale: CGFloat = 0.0 // Use scale factor of main screen
-        UIGraphicsBeginImageContextWithOptions(sizeChange, !hasAlpha, scale)
-        self.draw(in: CGRect(origin: CGPoint.zero, size: sizeChange))
-        let scaledImage = UIGraphicsGetImageFromCurrentImageContext()
-        return scaledImage!
-    }
-}
