+++ iOS/Severenity/Views/Navigation Bar/NavigationBarView.swift
@@ -1,0 +1,17 @@
+//
+//  NavigationBarView.swift
+//  Severenity
+//
+//  Created by Yura Yasinskyy on 28.10.16.
+//  Copyright © 2016 severenity. All rights reserved.
+//
+
+import UIKit
+
+class NavigationBarView: UIView {
+
+    @IBOutlet weak var userPicture: UIImageView!
+    @IBOutlet weak var userName: UILabel!
+    @IBOutlet weak var logOutButton: UIButton!
+    
+}
